refactor: extract TestServer to common test module and run all tests in CI
- Created tests/common/mod.rs with shared TestServer implementation
- Both integration tests now import from common module
- Eliminates code duplication between test files
- Updated CI workflow to run all integration tests (--tests) instead of just database_test
